# 콘솔로그 삭제

+++ src/pages/oauth/index.js
@@ -57,21 +57,19 @@ const OAuth = ({ match, location }) => {
 
     const Register =async(email,name,register_type)=>{
         try{
-            const res = await socialRegister(email,name,register_type);
-            console.log('소셜 회원가입');
-            console.log(res);
-
-            if(res.data.msg==='존재하는 이메일 주소로 가입을 시도하셔서 가입에 실패하셨습니다.'){
+            if(register_type==='already'){
                 history.replace('/');
-                openModal('회원가입 실패', res.data.msg);
+                openModal('회원가입 실패', '존재하는 이메일 주소로 가입을 시도하셔서 가입에 실패하셨습니다.');
             }
-            else if(res.data.access_token){
-                dispatch(get_user_info(res.data.access_token));
-                localStorage.setItem('access_token', res.data.access_token);
-                initStore();
-                history.replace('/');
+            else{
+                const res = await socialRegister(email,name,register_type);
+                if(res.data.access_token){
+                    dispatch(get_user_info(res.data.access_token));
+                    localStorage.setItem('access_token', res.data.access_token);
+                    initStore();
+                    history.replace('/');
+                }
             }
-            history.replace('/');
         } catch (e) {
             history.replace('/error');
         }
